Move the Camera class to its own file.
