+++ pyvista/plotting/__init__.py
@@ -6,12 +6,13 @@ from .export_vtkjs import export_plotter_vtkjs, get_vtkjs_url
 from .helpers import plot, plot_arrows, plot_compare_four, plot_itk
 from .itkplotter import PlotterITK
 from .plotting import BasePlotter, Plotter, close_all
-from .renderer import CameraPosition, Camera, Renderer, scale_point
+from .renderer import CameraPosition, Renderer, scale_point
 from .theme import (DEFAULT_THEME, FONT_KEYS, MAX_N_COLOR_BARS,
                     parse_color, parse_font_family, rcParams, set_plot_theme)
 from .tools import (create_axes_marker, create_axes_orientation_box,
                     opacity_transfer_function, system_supports_plotting)
 from .widgets import WidgetHelper
+from .camera import Camera
 
 
 class QtDeprecationError(Exception):
